reorg: put all new code in int_set_par
